Large rewrite.  All remote admin and console access has been moved to MgmNode, so we no longer have plaintext system account credentials on our frontend.  Oar functionality, except for nuke, is operational.  The console needs to be rewritten to remove dependence on the misbehaving halcyon rest console.

+++ src/mgm/routes/TaskHandler.ts
@@ -6,6 +6,7 @@ import { Job } from '../Job';
 import { MGM } from '../MGM';
 import { UUIDString } from '../../halcyon/UUID';
 import { Region } from '../Region';
+import { Host } from '../Host';
 
 import fs = require("fs");
 import * as multer from 'multer';
@@ -66,13 +67,17 @@ export function TaskHandler(mgm: MGM, uploadDir: string): express.Router {
   router.post('/delete/:id', MGM.isUser, (req, res) => {
     let taskID = parseInt(req.params.id);
 
-    mgm.getJob(taskID).then( (j: Job) => {
+    mgm.getJob(taskID).then((j: Job) => {
       let datum = JSON.parse(j.data);
-      if( datum.File && datum.File !== ''){
-        fs.unlink( datum.File);
+      if (datum.File && datum.File !== '') {
+        fs.exists(datum.File, (exists) => {
+          if (exists) {
+            fs.unlink(datum.File);
+          }
+        });
       }
       return mgm.deleteJob(j);
-    }).then( () => {
+    }).then(() => {
       res.send(JSON.stringify({ Success: true }));
     }).catch((err: Error) => {
       res.send(JSON.stringify({ Success: false, Message: err.message }));
@@ -80,7 +85,41 @@ export function TaskHandler(mgm: MGM, uploadDir: string): express.Router {
   });
 
   router.post('/saveOar/:uuid', MGM.isUser, (req, res) => {
-    res.send(JSON.stringify({ Success: false, Message: 'Not Implemented' }));
+    let regionID = new UUIDString(req.params.uuid);
+    let user = new UUIDString(req.cookies['uuid']);
+
+    let region: Region;
+    let host: Host;
+
+    mgm.getRegion(regionID).then((r: Region) => {
+      console.log('User ' + user + ' requesting save oar for region ' + regionID);
+      if (!r.isRunning) {
+        throw new Error('Region is not running');
+      }
+      region = r;
+      return mgm.getHost(r.slaveAddress);
+    }).then((h: Host) => {
+      host = h;
+
+      let j: Job = {
+        id: 0,
+        timestamp: '',
+        type: 'save_oar',
+        user: user,
+        data: JSON.stringify({
+          Status: 'Pending...',
+          Region: regionID.toString()
+        })
+      };
+
+      return mgm.insertJob(j);
+    }).then((j: Job) => {
+      return mgm.saveOar(region, host, j);
+    }).then(() => {
+      res.send(JSON.stringify({ Success: true }));
+    }).catch((err: Error) => {
+      res.send(JSON.stringify({ Success: false, Message: err.message }));
+    });
   });
 
   router.post('/nukeContent/:uuid', MGM.isUser, (req, res) => {
@@ -103,36 +142,99 @@ export function TaskHandler(mgm: MGM, uploadDir: string): express.Router {
     res.send(JSON.stringify({ Success: false, Message: 'Not Implemented' }));
   });
 
-  router.get('/ready', (req, res) => {
-    res.send('MGM');
+  router.get('/ready/:id', (req, res) => {
+    let jobID = parseInt(req.params.id);
+
+    mgm.getJob(jobID).then( (j: Job) => {
+      switch(j.type){
+        case 'save_oar':
+          let user = new UUIDString(req.cookies['uuid']);
+          if(j.user.toString() !== user.toString()){
+            throw new Error('Permission Denied');
+          }
+          let datum = JSON.parse(j.data);
+          res.setHeader('Content-Disposition', 'attachment; filename="'+datum.FileName+'.oar"');
+          res.setHeader('Content-Type', 'application/octet-stream');
+          res.sendFile(datum.File);
+          break;
+        case 'load_oar':
+          let remoteIP: string = req.ip.split(':').pop();
+          return mgm.getHost(remoteIP).then( (h: Host) => {
+            //valid host, serve the file
+
+          });
+      }
+
+    }).catch( (err) => {
+      console.log('An error occurred sending a file to a user: ' + err);
+    });
   });
 
-  router.get('/report', (req, res) => {
-    res.send('MGM');
+  router.post('/report/:id', (req, res) => {
+    let taskID = parseInt(req.params.id);
+    let remoteIP: string = req.ip.split(':').pop();
+
+    mgm.getHost(remoteIP).then((h: Host) => {
+      return mgm.getJob(taskID);
+    }).then( (j: Job) => {
+      let datum = JSON.parse(j.data);
+      datum.Status = req.body.Status;
+      j.data = JSON.stringify(datum);
+      return mgm.updateJob(j);
+    }).then( () => {
+      res.send('OK');
+    }).catch( (err) => {
+      console.log(err);
+    });
   });
 
-  router.post('/upload/:id', multer({ dest: uploadDir}).single('file'), (req, res) => {
+  router.post('/upload/:id', multer({ dest: uploadDir }).single('file'), (req, res) => {
     let taskID = parseInt(req.params.id);
 
     console.log('upload file received for job ' + taskID);
 
-    mgm.getJob(taskID).then( (j: Job) => {
-      switch(j.type){
+    mgm.getJob(taskID).then((j: Job) => {
+      switch (j.type) {
+        case 'save_oar':
+          let remoteIP: string = req.ip.split(':').pop();
+          return mgm.getHost(remoteIP).then((h: Host) => {
+            //host is valid
+            let datum = JSON.parse(j.data);
+            datum.Status = 'Done';
+            datum.File = req.file.path;
+            datum.FileName = req.file.originalname;
+            datum.Size = req.file.size;
+            j.data = JSON.stringify(datum);
+            return mgm.updateJob(j);
+          }).then( () => {});
         case 'load_oar':
+          let user = new UUIDString(req.cookies['uuid']);
+          if(user.toString() !== j.user.toString()){
+            throw new Error('Permission Denied');
+          }
           let datum = JSON.parse(j.data);
           datum.Status = "Loading";
           datum.File = req.file.path;
           j.data = JSON.stringify(datum);
-          return mgm.updateJob(j);
+          let region: Region;
+          return mgm.updateJob(j).then( (j: Job) =>{
+            return mgm.getRegion(datum.Region);
+          }).then( (r: Region) => {
+            region = r;
+            if(! r.isRunning){
+              throw new Error('Region is not running');
+            }
+            return mgm.getHost(r.slaveAddress);
+          }).then( (h: Host) => {
+            return mgm.loadOar(region, h, j);
+          })
         default:
           throw new Error('invalid upload for job type: ' + j.type);
       }
-    }).then( (j: Job) => {
-      //do oar loading magics.
-      return mgm.doJob(j);
-    }).then( () => {
+    }).then(() => {
       res.send(JSON.stringify({ Success: true }));
     }).catch((err: Error) => {
+      console.log(err);
       res.send(JSON.stringify({ Success: false, Message: err.message }));
     });
   });
